design & seederhome & create function plan is done

+++ database/migrations/2022_04_26_113954_create_plans_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->json('description');
-            $table->integer('prices');
+            $table->decimal('prices');
             $table->timestamps();
         });
     }
